+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][Keep a Changelog] and this project adheres to [Semantic Versioning][Semantic Versioning].
 
+## [Unreleased]
+
+- Added: Create parent directory before rsync'ing.
+
 ## [1] - 2021-02-28
 
-- Added: Support rsync uploads
+- Added: Support rsync uploads.
 - Added: `rsync` and `ssh` flags can be changed via configuration.